Refactor: Split monolithic main.py into APIRouters
