+++ backend/main.py
@@ -35,186 +35,23 @@ from dotenv import load_dotenv
 env_path = Path(__file__).parent / '.env'
 load_dotenv(dotenv_path=env_path)
 
-# Initialize Supabase Client (Service Role for backend bypass)
-try:
-    from supabase import create_client, Client
-    url = os.environ.get("SUPABASE_URL")
-    key = os.environ.get("SUPABASE_SERVICE_KEY")
-    if not url or not key:
-        print("[ERROR] SUPABASE_URL or SUPABASE_SERVICE_KEY not set in backend/.env")
-        supabase = None
-    else:
-        supabase = create_client(url, key)
-except (ImportError, Exception) as e:
-    print(f"[WARNING] Supabase initialization failed: {e}")
-    supabase = None
-    Client = None
-
 # Ensure project root is on path for imports
+import sys
+import os
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from backend.services.classifier_service import ClassifierService
-from backend.services.classifier_v2 import classifier_v2
-from backend.services.classifier_v3 import classifier_v3 # V3 Power Model
-from backend.services.ner_service import NERService
-from backend.services.duplicate_service import DuplicateService
-from backend.services.rag_service import RagService
+from fastapi.middleware.cors import CORSMiddleware
+from fastapi.responses import HTMLResponse, JSONResponse
+from fastapi.encoders import jsonable_encoder
+from slowapi.errors import RateLimitExceeded
+from slowapi import _rate_limit_exceeded_handler
 
+from backend.dependencies import (
+    classifier_service, ner_service, duplicate_service, rag_service, gemini_service, supabase, limiter
+)
+from backend.models import HealthResponse, ReadinessResponse
+from backend.routers import auth, tickets, ai
 
-# ---------------------------------------------------------------------------
-# Request / Response models
-# ---------------------------------------------------------------------------
-def get_system_settings(company_id: str) -> dict:
-    defaults = {
-        "ai_confidence_threshold": 0.80,
-        "duplicate_sensitivity": 0.85,
-        "enable_auto_resolve": False
-    }
-    if not supabase or not company_id:
-        return defaults
-    try:
-        res = supabase.table("system_settings").select(
-            "ai_confidence_threshold, duplicate_sensitivity, enable_auto_resolve"
-        ).eq("company_id", company_id).single().execute()
-        if res.data:
-            return {**defaults, **res.data}
-    except Exception as e:
-        print(f"[WARNING] Could not fetch system_settings for company_id={company_id}: {e}")
-    return defaults
-class TicketRequest(BaseModel):
-    text: str
-    image_base64: str = ""
-    image_text: str = "" # Keep for backward compatibility
-    user_id: str | None = None
-    company: str | None = None
-    image_url: str | None = None
-    confidence_threshold: float = 0.20
-    duplicate_sensitivity: float = 0.85
-
-class TicketSaveRequest(BaseModel):
-    user_id: str
-    subject: str
-    description: str
-    category: str
-    subcategory: str
-    priority: str
-    assigned_team: str
-    status: str
-    auto_resolve: bool
-    is_duplicate: bool
-    confidence: float
-    image_url: str | None = None
-    company: str | None = None
-    company_id: str | None = None
-    sla_breach_at: str
-    metadata: dict
-    entities: list = []
-    solution_steps: list = []
-    ocr_text: str = ""
-    needs_review: bool = False
-    routing_confidence: float
-
-
-class DuplicateInfo(BaseModel):
-    is_duplicate: bool
-    duplicate_ticket_id: str | None = None
-    similarity: float = 0.0
-
-
-class EntityInfo(BaseModel):
-    text: str
-    label: str
-    confidence: float
-
-
-class TicketResponse(BaseModel):
-    id: str | int | None = None
-    ticket_id: str | None = None
-    summary: str
-    category: str
-    subcategory: str
-    priority: str
-    auto_resolve: bool
-    assigned_team: str
-    entities: list[EntityInfo]
-    duplicate_ticket: DuplicateInfo
-    confidence: float
-    needs_review: bool = False
-    reasoning: str = ""
-    decision_factors: list[str] = []
-    image_description: str = ""
-    ocr_text: str = ""
-    image_url: str | None = None
-    highlights: list[str] = []
-    timeline: dict = {} # Map of step_name: timestamp
-    env_metadata: dict = {} # IP, Hostname, Browser/OS
-    sla_breach_at: str | None = None
-    version: str = "2.1.0-Neural-Diagnostic"
-
-
-# --- Persistence Models ---
-class Message(BaseModel):
-    sender: str
-    message: str
-    timestamp: str
-
-
-class TicketRecord(BaseModel):
-    ticket_id: str
-    owner_id: str
-    summary: str
-    category: str
-    subcategory: str
-    priority: str
-    status: str
-    assigned_team: str
-    created_at: str
-    updated_at: str | None = None
-    last_user_viewed_at: str | None = None
-    messages: list[Message] = []
-    metadata: dict = {}
-    timeline: dict = {} # Milestones: created, analyzed, triaged, routed, in_progress, resolved
-
-
-# --- In-Memory Database (to be replaced with SQL later) ---
-TICKETS_DB: list[TicketRecord] = []
-
-
-class HealthResponse(BaseModel):
-    status: str
-    classifier_loaded: bool
-    ner_loaded: bool
-
-
-class ReadinessResponse(BaseModel):
-    status: str
-    checks: dict[str, bool]
-
-
-# ---------------------------------------------------------------------------
-# Service singletons
-# ---------------------------------------------------------------------------
-classifier_service = ClassifierService()
-ner_service = NERService()
-duplicate_service = DuplicateService()
-rag_service = RagService()
-
-try:
-    from backend.services.gemini_service import GeminiService
-    gemini_service = GeminiService()
-except ImportError:
-    gemini_service = None
-
-try:
-    from backend.services.ocr_service import OCRService
-    ocr_service = OCRService()
-except ImportError:
-    ocr_service = None
-
-
-# ---------------------------------------------------------------------------
-# Lifespan (startup / shutdown)
-# ---------------------------------------------------------------------------
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     """Load all models at startup."""
@@ -259,9 +96,7 @@ async def lifespan(app: FastAPI):
     print("[Shutdown] Cleaning up ...")
 
 
-# ---------------------------------------------------------------------------
-# App
-# ---------------------------------------------------------------------------
+
 app = FastAPI(
     title="AI Helpdesk Backend",
     description="Ticket classification, entity extraction, and duplicate detection",
@@ -269,12 +104,9 @@ app = FastAPI(
     lifespan=lifespan,
 )
 
-# Rate limiter — 10 AI requests per minute per IP (free tier protection)
-limiter = Limiter(key_func=get_remote_address)
 app.state.limiter = limiter
 app.add_exception_handler(RateLimitExceeded, _rate_limit_exceeded_handler)
 
-# CORS — load from environment variable with fallback
 allowed_origins_env = os.environ.get("ALLOWED_ORIGINS")
 if allowed_origins_env:
     origins = [origin.strip() for origin in allowed_origins_env.split(",") if origin.strip()]
@@ -293,10 +125,9 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-
-# ---------------------------------------------------------------------------
-# Root & Health check
-# ---------------------------------------------------------------------------
+app.include_router(auth.router)
+app.include_router(tickets.router)
+app.include_router(ai.router)
 @app.get("/", response_class=HTMLResponse)
 async def root():
     return """
@@ -419,799 +250,4 @@ async def readiness_check():
         status_code=503,
         content=jsonable_encoder(ReadinessResponse(status="not_ready", checks=checks)),
     )
-
-
-class TroubleshootRequest(BaseModel):
-    text: str
-    category: str
-    history: list[dict] = []
-
-class TroubleshootResponse(BaseModel):
-    step_text: str
-    options: list[str]
-    is_final: bool
-
-@app.post("/ai/troubleshoot", response_model=TroubleshootResponse)
-async def troubleshoot(request: TroubleshootRequest):
-    """Get dynamic troubleshooting steps from Gemini."""
-    if not gemini_service or not gemini_service._initialized:
-        return TroubleshootResponse(
-            step_text="AI Troubleshooting is currently unavailable.",
-            options=["Continue to tracking"],
-            is_final=True
-        )
-    
-    result = gemini_service.get_troubleshooting_step(
-        request.text,
-        request.history,
-        request.category
-    )
-    return TroubleshootResponse(**result)
-
-
-class BugReportAnalysisRequest(BaseModel):
-    bug_title: str
-    description: str
-    steps_to_reproduce: str = ""
-    console_errors: list[str] = []
-
-class BugReportAnalysisResponse(BaseModel):
-    probable_cause: str
-
-@app.post("/ai/analyze_bug", response_model=BugReportAnalysisResponse)
-async def analyze_bug(request: BugReportAnalysisRequest):
-    """Analyze a bug report using Gemini to generate a Probable Cause."""
-    if not gemini_service or not gemini_service._initialized:
-        return BugReportAnalysisResponse(
-            probable_cause="AI Diagnostics are currently unavailable."
-        )
-    
-    cause = gemini_service.analyze_bug_report(
-        request.bug_title,
-        request.description,
-        request.steps_to_reproduce,
-        request.console_errors
-    )
-    return BugReportAnalysisResponse(probable_cause=cause)
-
-
-# ---------------------------------------------------------------------------
-# Admin Correction Logging endpoint
-# ---------------------------------------------------------------------------
-CORRECTIONS_LOG_PATH = Path(__file__).parent / "data" / "corrections_log.json"
-
-@app.post("/ai/log_correction")
-async def log_correction(raw_request: Request):
-    """Log an admin correction when the AI prediction differs from the human decision."""
-    try:
-        body = await raw_request.json()
-    except Exception as e:
-        print(f"[CORRECTION ERROR] Could not parse request body: {e}")
-        return {"status": "error", "message": "Invalid JSON body"}
-
-    print(f"[CORRECTION RECEIVED] Payload keys: {list(body.keys())}")
-
-    ticket_id = str(body.get("ticket_id", "unknown"))
-    original_text = str(body.get("original_text", ""))
-    ocr_text = str(body.get("ocr_text", ""))
-    confidence = float(body.get("confidence") or 0.0)
-    original_prediction = body.get("original_prediction") or {}
-    corrected_prediction = body.get("corrected_prediction") or {}
-
-    # Only log if something actually changed
-    changed_fields = [
-        field for field in ["category", "subcategory", "priority", "assigned_team"]
-        if original_prediction.get(field) != corrected_prediction.get(field)
-    ]
-
-    if not changed_fields:
-        return {"status": "no_change", "message": "Prediction matches correction, nothing logged."}
-
-    entry = {
-        "ticket_id": ticket_id,
-        "original_text": original_text,
-        "ocr_text": ocr_text,
-        "original_prediction": original_prediction,
-        "corrected_prediction": corrected_prediction,
-        "changed_fields": changed_fields,
-        "confidence": confidence,
-        "timestamp": datetime.datetime.utcnow().isoformat() + "Z"
-    }
-
-    try:
-        if CORRECTIONS_LOG_PATH.exists() and CORRECTIONS_LOG_PATH.stat().st_size > 2:
-            with open(CORRECTIONS_LOG_PATH, "r", encoding="utf-8") as f:
-                logs = json.load(f)
-        else:
-            logs = []
-
-        logs.append(entry)
-
-        with open(CORRECTIONS_LOG_PATH, "w", encoding="utf-8") as f:
-            json.dump(logs, f, indent=2)
-
-        print(f"[CORRECTION SAVED] Ticket ID: {ticket_id} | Changed: {changed_fields}")
-        return {"status": "saved", "changed_fields": changed_fields}
-
-    except Exception as e:
-        print(f"[CORRECTION ERROR] Could not save: {e}")
-        return {"status": "error", "message": str(e)}
-
-
-# ---------------------------------------------------------------------------
-# Ticket operations (Now via Supabase)
-# ---------------------------------------------------------------------------
-@app.get("/tickets")
-async def get_tickets(company_id: str | None = None):
-    """Fetch persistent tickets from Supabase."""
-    if not supabase:
-        raise HTTPException(status_code=500, detail="Database connection not initialized")
-    
-    query = supabase.table("tickets").select("*").order("created_at", desc=True)
-    if company_id:
-        query = query.eq("company_id", company_id)
-        
-    res = query.execute()
-    return res.data
-
-@app.post("/tickets/save")
-async def save_ticket(request_body: TicketSaveRequest):
-    """
-    OFFICIAL PERSISTENCE: Saves the analyzed ticket to Supabase.
-    This is called AFTER the user confirms the analysis results.
-    """
-    if not supabase:
-        raise HTTPException(status_code=500, detail="Supabase connection not initialized.")
-
-    logger = logging.getLogger(__name__)
-    try:
-        final_data = request_body.dict()
-
-        # Resolve tenant linkage from user profile with authorization validation.
-        profile = {}
-        if request_body.user_id:
-            try:
-                profile_res = (
-                    supabase.table("profiles")
-                    .select("company_id, company")
-                    .eq("id", request_body.user_id)
-                    .single()
-                    .execute()
-                )
-                profile = profile_res.data or {}
-                if not profile:
-                    raise HTTPException(status_code=404, detail="User profile not found")
-            except HTTPException:
-                raise
-            except Exception as profile_error:
-                user_hash = hashlib.sha256(str(request_body.user_id).encode()).hexdigest()[:8]
-                logger.error(f"Tenant resolution error for user {user_hash}: {profile_error}")
-                raise HTTPException(status_code=503, detail="Failed to resolve tenant linkage") from profile_error
-
-        # Validate tenant consistency and authorization.
-        profile_company_id = profile.get("company_id")
-        if final_data.get("company_id"):
-            # User provided company_id: verify it matches their profile.
-            if profile_company_id and final_data["company_id"] != profile_company_id:
-                user_hash = hashlib.sha256(str(request_body.user_id).encode()).hexdigest()[:8]
-                logger.warning(f"Tenant mismatch: user {user_hash} attempted {final_data['company_id']}, assigned to {profile_company_id}")
-                raise HTTPException(status_code=403, detail="User not authorized for this tenant")
-        elif profile_company_id:
-            # Backfill company_id from profile.
-            final_data["company_id"] = profile_company_id
-        elif request_body.user_id:
-            # User has no tenant assignment.
-            raise HTTPException(status_code=400, detail="User has no tenant assignment")
-
-        # Backfill company name if missing.
-        if not final_data.get("company") and profile.get("company"):
-            final_data["company"] = profile["company"]
-
-        user_hash = hashlib.sha256(str(request_body.user_id).encode()).hexdigest()[:8]
-        logger.info(f"Tenant linkage: user_hash={user_hash}, company_id={final_data.get('company_id')}")
-
-
-        res = supabase.table("tickets").insert(final_data).execute()
-        
-        if not res.data:
-            raise Exception("Failed to insert ticket into database.")
-            
-        ticket_id = res.data[0]["id"]
-
-        duplicate_indexed = True
-        duplicate_index_warning = None
-        description_text = (request_body.description or "").strip()
-        subject_text = (request_body.subject or "").strip()
-        duplicate_text = description_text or subject_text
-        if duplicate_text:
-            try:
-                duplicate_service.add_ticket(str(ticket_id), duplicate_text)
-            except Exception as index_error:
-                duplicate_indexed = False
-                duplicate_index_warning = "Duplicate index update failed."
-                print(f"[WARNING] {duplicate_index_warning} ticket_id={ticket_id} error={index_error}")
-        else:
-            duplicate_indexed = False
-            duplicate_index_warning = "Duplicate index update skipped: no description or subject text was provided."
-            print(f"[WARNING] {duplicate_index_warning}")
-        
-        # Add initial system diagnostic message
-        msg = "Our Neural Engine has successfully triaged your issue and routed it to the designated team."
-        if final_data["auto_resolve"]:
-            msg = "AI Auto-Resolution active: A verified solution has been identified. Please review the attached resolution steps."
-
-        supabase.table("ticket_messages").insert({
-            "ticket_id": ticket_id,
-            "sender_id": "00000000-0000-0000-0000-000000000000", # System ID
-            "sender_name": "AI Assistant",
-            "sender_role": "admin",
-            "message": msg
-        }).execute()
-        
-        response = {"status": "success", "ticket_id": ticket_id, "duplicate_indexed": duplicate_indexed}
-        if duplicate_index_warning:
-            response["duplicate_index_warning"] = duplicate_index_warning
-        return response
-
-    except Exception as e:
-        traceback.print_exc()
-        raise HTTPException(status_code=500, detail=str(e))
-
-@app.get("/tickets/{ticket_id}")
-async def get_ticket_by_id(ticket_id: str):
-    """Fetch single persistent ticket."""
-    if not supabase:
-        raise HTTPException(status_code=500, detail="Database connection not initialized")
-    
-    res = supabase.table("tickets").select("*").eq("id", ticket_id).single().execute()
-    if not res.data:
-        raise HTTPException(status_code=404, detail="Ticket not found")
-    return res.data
-
-
-@app.post("/tickets", response_model=TicketRecord)
-async def create_ticket(ticket: TicketRecord):
-    """Save a new ticket into the system."""
-    # Check for duplicates before adding
-    existing = next((t for t in TICKETS_DB if t.ticket_id == ticket.ticket_id), None)
-    if existing:
-        return existing
-        
-    TICKETS_DB.append(ticket)
-    print(f"[DB] Ticket #{ticket.ticket_id} created for user {ticket.owner_id}")
-    return ticket
-
-
-@app.patch("/tickets/{ticket_id}", response_model=TicketRecord)
-async def update_ticket(ticket_id: str, updates: dict):
-    """Partially update a ticket's fields (e.g., status, viewed_at)."""
-    for i, ticket in enumerate(TICKETS_DB):
-        if str(ticket.ticket_id) == str(ticket_id):
-            # Convert to dict, update, then back to model
-            ticket_dict = ticket.dict()
-            ticket_dict.update(updates)
-            updated_ticket = TicketRecord(**ticket_dict)
-            TICKETS_DB[i] = updated_ticket
-            return updated_ticket
-    
-    raise HTTPException(status_code=404, detail="Ticket not found")
-
-
-# ---------------------------------------------------------------------------
-# Main AI Analyzer endpoint
-# ---------------------------------------------------------------------------
-@app.post("/ai/analyze_ticket", response_model=TicketResponse)
-@limiter.limit("10/minute")
-async def analyze_ticket(request_body: TicketRequest, request: Request):
-    """
-    Main endpoint for analyzing a new ticket using the cascade of local AI models.
-    """
-    text = request_body.text
-
-    settings = get_system_settings(request_body.company)
-    confidence_threshold = settings["ai_confidence_threshold"]
-    duplicate_sensitivity = settings["duplicate_sensitivity"]
-    enable_auto_resolve = settings["enable_auto_resolve"]
-    
-    # Grab client metadata
-    client_ip = request.client.host if request.client else "unknown"
-    user_agent = request.headers.get("user-agent", "unknown")
-    origin_host = request.headers.get("origin", "unknown")
-    
-    env_metadata = {
-        "ip": client_ip,
-        "user_agent": user_agent,
-        "origin": origin_host
-    }
-
-    # --- Layer 1: Local OCR (CPU, no API required) ---
-    local_ocr_text = ""
-    if request_body.image_base64 and ocr_service:
-        print("[AI] Extracting text via local OCR...")
-        local_ocr_text = ocr_service.extract_text(request_body.image_base64)
-        if local_ocr_text:
-            text = f"{text} {local_ocr_text}".strip()
-            print(f"[AI] OCR added {len(local_ocr_text)} chars to context.")
-
-    # Initalize Timeline
-    return await analyze_only(request_body)
-
-@app.post("/ai/analyze")
-async def analyze_only(request_body: TicketRequest):
-    """
-    PERFORMANCE UPGRADE: AI Analysis phase only. 
-    Does NOT persist to DB. This allows the user to review the analysis 
-    and duplicate check before committing to a ticket creation.
-    """
-    text = request_body.text
-    print(f"[AI] Starting Analysis (READ-ONLY) for: {text[:50]}...") 
-    settings = get_system_settings(request_body.company)
-    confidence_threshold = settings["ai_confidence_threshold"]
-    duplicate_sensitivity = settings["duplicate_sensitivity"]
-    enable_auto_resolve = settings["enable_auto_resolve"]
-    
-    # --- Context & Environment ---
-    import datetime
-    def get_now_ist():
-        return datetime.datetime.utcnow().isoformat() + "Z"
-
-    env_metadata = {
-        "timestamp": get_now_ist(),
-        "model_version": "3.0.0-PRO",
-        "api_endpoint": "/ai/analyze"
-    }
-    
-    timeline = {"received": get_now_ist()}
-
-    # --- Vision Logic (OCR Awareness) ---
-    gemini_analysis = {
-        "ocr_text": request_body.image_text or "",
-        "image_description": ""
-    }
-    
-    if request_body.image_base64 and not gemini_analysis["ocr_text"]:
-        try:
-            print("[AI] Detecting visual context via Gemini...")
-            vision_result = gemini_service.analyze_image(request_body.image_base64, text)
-            gemini_analysis.update(vision_result)
-        except Exception as e:
-            print(f"[VISION ERROR] {e}")
-
-    summary = text[:100] + ("…" if len(text) > 100 else "") 
-
-    # --- Classification ---
-    try:
-        classification_v3_res = classifier_v3.predict(text)
-        if "error" in classification_v3_res:
-            # Fallback to V1
-            classification = classifier_service.predict(text)
-        else:
-            # Parse V3 output
-            cat = classification_v3_res.get("Category", {}).get("prediction", "Unknown")
-            sub = classification_v3_res.get("Subcategory", {}).get("prediction", "Unknown")
-            pri = classification_v3_res.get("priority", {}).get("prediction", "Medium")
-            conf = classification_v3_res.get("Category", {}).get("confidence", 0.0)
-            
-            from backend.services.classifier_service import TEAM_MAP, AUTO_RESOLVE_SUBS
-            assigned_team = TEAM_MAP.get(cat, "General Support")
-            auto_resolve = sub in AUTO_RESOLVE_SUBS
-            
-            classification = {
-                "category": cat,
-                "subcategory": sub,
-                "priority": pri,
-                "auto_resolve": auto_resolve,
-                "assigned_team": assigned_team,
-                "confidence": float(conf)
-            }
-    except Exception as e:
-        traceback.print_exc()
-        classification = {
-            "category": "Unknown", "subcategory": "Unknown", "priority": "Medium",
-            "auto_resolve": False, "assigned_team": "General Support", "confidence": 0.0,
-        }
-
-    timeline["ai_analyzed"] = get_now_ist()
-    timeline["triaged"] = get_now_ist()
-
-    # --- NER ---
-    try:
-        entities = ner_service.extract_entities(text)
-    except Exception:
-        entities = []
-    
-    timeline["metadata_harvested"] = get_now_ist()
-
-    # --- Duplicate detection ---
-    try:
-        dup_result = duplicate_service.check_duplicate(text, threshold=duplicate_sensitivity)
-    except Exception:
-        dup_result = {"is_duplicate": False, "duplicate_ticket_id": None, "similarity": 0.0}
-
-    # --- RAG Knowledge Base Check ---
-    rag_match = None
-    try:
-        rag_match = rag_service.search_knowledge_base(text, threshold=0.85)
-        if rag_match:
-            classification["auto_resolve"] = True
-            classification["assigned_team"] = "Auto-Resolve AI"
-            classification["confidence"] = max(classification["confidence"], float(rag_match["similarity"]))
-            print(f"[RAG SUCCESS] Found solution for: '{rag_match['title']}'")
-    except Exception as e:
-        print(f"[RAG ERROR] {e}")
-
-    # --- Reasoning ---
-    decision_factors = []
-    if classification["confidence"] > confidence_threshold:
-        decision_factors.append(f"High confidence match for '{classification['subcategory']}'")
-    if entities:
-        decision_factors.append(f"Detected entities: {', '.join([e['text'] for e in entities[:2]])}")
-    if dup_result["is_duplicate"]:
-        decision_factors.append(f"Found similar incident ({int(dup_result['similarity']*100)}%)")
-    if rag_match:
-        decision_factors.append(f"Found solution article: '{rag_match['title']}'")
-
-    reasoning = f"Categorized as '{classification['category']}' - {classification['subcategory']}."
-    if (
-        enable_auto_resolve
-        and classification["confidence"] >= confidence_threshold
-        and classification["auto_resolve"]
-    ):
-        classification["auto_resolve"] = True
-    else:
-        classification["auto_resolve"] = False
-    if classification["auto_resolve"]:
-        reasoning += " Flagged for AI auto-resolution via Knowledge Base." if rag_match else " Flagged for auto-resolution."
-    
-    timeline["routed"] = get_now_ist()
-    
-    # --- Gemini Summary ---
-    if gemini_service and gemini_service._initialized:
-        summary = gemini_service.get_summary(text)
-    
-    # Convert priority to SLA breached timestamp (for preview)
-    hours_map = {"Critical": 2, "High": 8, "Medium": 24, "Low": 72}
-    sla_hours = hours_map.get(classification["priority"], 72)
-    sla_breach_dt = datetime.datetime.utcnow() + datetime.timedelta(hours=sla_hours)
-
-    return TicketResponse(
-        ticket_id=str(uuid.uuid4()), # Temporary ID
-        summary=summary,
-        category=classification["category"],
-        subcategory=classification["subcategory"],
-        priority=classification["priority"],
-        auto_resolve=classification["auto_resolve"],
-        assigned_team=classification["assigned_team"],
-        entities=[EntityInfo(**e) for e in entities],
-        duplicate_ticket=DuplicateInfo(**dup_result),
-        confidence=classification["confidence"],
-        needs_review=classification["confidence"] < confidence_threshold,
-        reasoning=reasoning,
-        decision_factors=decision_factors,
-        image_description=gemini_analysis["image_description"],
-        ocr_text=gemini_analysis["ocr_text"],
-        image_url=request_body.image_url,
-        highlights=entities, # Use entities as highlights for now
-        timeline=timeline,
-        env_metadata=env_metadata,
-        sla_breach_at=sla_breach_dt.isoformat() + "Z"
-    )
-
-@app.post("/ai/analyze_stream")
-async def analyze_stream(request_body: TicketRequest):
-    """
-    REAL-TIME SSE ENDPOINT: Streams the AI progress to the frontend dynamically.
-    """
-    import datetime
-    def get_now_ist():
-        return datetime.datetime.utcnow().isoformat() + "Z"
-
-    async def event_generator():
-        text = request_body.text
-        env_metadata = {
-            "timestamp": get_now_ist(),
-            "model_version": "3.0.0-PRO",
-            "api_endpoint": "/ai/analyze_stream"
-        }
-        timeline = {"received": get_now_ist()} 
-        settings = get_system_settings(request_body.company)
-        confidence_threshold = settings["ai_confidence_threshold"]
-        duplicate_sensitivity = settings["duplicate_sensitivity"]
-        enable_auto_resolve = settings["enable_auto_resolve"]
-
-        # 1. Reading
-        yield f"data: {json.dumps({'step': 'Reading your message', 'status': 'in_progress'})}\n\n"
-        await asyncio.sleep(0.5)
-
-        gemini_analysis = {"ocr_text": request_body.image_text or "", "image_description": ""}
-        if request_body.image_base64 and not gemini_analysis["ocr_text"]:
-            try:
-                vision_result = gemini_service.analyze_image(request_body.image_base64, text)
-                gemini_analysis.update(vision_result)
-            except Exception as e:
-                pass
-
-        summary = text[:100] + ("…" if len(text) > 100 else "") 
-
-        # 2. NER
-        yield f"data: {json.dumps({'step': 'Extracting technical entities', 'status': 'in_progress'})}\n\n"
-        await asyncio.sleep(0.2)
-        try:
-            entities = ner_service.extract_entities(text)
-        except Exception:
-            entities = []
-        timeline["metadata_harvested"] = get_now_ist()
-
-        # 3. Classification
-        yield f"data: {json.dumps({'step': 'Detecting category and priority', 'status': 'in_progress'})}\n\n"
-        await asyncio.sleep(0.2)
-        try:
-            classification_v3_res = classifier_v3.predict(text)
-            if "error" in classification_v3_res:
-                classification = classifier_service.predict(text)
-            else:
-                cat = classification_v3_res.get("Category", {}).get("prediction", "Unknown")
-                sub = classification_v3_res.get("Subcategory", {}).get("prediction", "Unknown")
-                pri = classification_v3_res.get("priority", {}).get("prediction", "Medium")
-                conf = classification_v3_res.get("Category", {}).get("confidence", 0.0)
-                
-                from backend.services.classifier_service import TEAM_MAP, AUTO_RESOLVE_SUBS
-                assigned_team = TEAM_MAP.get(cat, "General Support")
-                auto_resolve = sub in AUTO_RESOLVE_SUBS
-                
-                classification = {
-                    "category": cat,
-                    "subcategory": sub,
-                    "priority": pri,
-                    "auto_resolve": auto_resolve,
-                    "assigned_team": assigned_team,
-                    "confidence": float(conf)
-                }
-        except Exception as e:
-            classification = {
-                "category": "Unknown", "subcategory": "Unknown", "priority": "Medium",
-                "auto_resolve": False, "assigned_team": "General Support", "confidence": 0.0,
-            }
-        timeline["ai_analyzed"] = get_now_ist()
-        timeline["triaged"] = get_now_ist()
-
-        # 4. Duplicates
-        yield f"data: {json.dumps({'step': 'Checking duplicate issues', 'status': 'in_progress'})}\n\n"
-        await asyncio.sleep(0.2)
-        try:
-            dup_result = duplicate_service.check_duplicate(text, threshold=duplicate_sensitivity)
-        except Exception:
-            dup_result = {"is_duplicate": False, "duplicate_ticket_id": None, "similarity": 0.0}
-
-        # 5. RAG / Solutions
-        yield f"data: {json.dumps({'step': 'Finding possible solutions', 'status': 'in_progress'})}\n\n"
-        await asyncio.sleep(0.2)
-        rag_match = None
-        try:
-            rag_match = rag_service.search_knowledge_base(text, threshold=0.85)
-            if rag_match:
-                classification["auto_resolve"] = True
-                classification["assigned_team"] = "Auto-Resolve AI"
-                classification["confidence"] = max(classification["confidence"], float(rag_match["similarity"]))
-        except Exception as e:
-            pass
-
-        decision_factors = []
-        if classification["confidence"] > confidence_threshold:
-            decision_factors.append(f"High confidence match for '{classification['subcategory']}'")
-        if entities:
-            decision_factors.append(f"Detected entities: {', '.join([e['text'] for e in entities[:2]])}")
-        if dup_result["is_duplicate"]:
-            decision_factors.append(f"Found similar incident ({int(dup_result['similarity']*100)}%)")
-        if rag_match:
-            decision_factors.append(f"Found solution article: '{rag_match['title']}'")
-
-        if not enable_auto_resolve:
-            classification["auto_resolve"] = False
-        reasoning = f"Categorized as '{classification['category']}' - {classification['subcategory']}."
-        if classification["auto_resolve"]:
-            reasoning += " Flagged for AI auto-resolution via Knowledge Base." if rag_match else " Flagged for auto-resolution."
-        
-        timeline["routed"] = get_now_ist()
-
-        if gemini_service and gemini_service._initialized:
-            summary = gemini_service.get_summary(text)
-        
-        hours_map = {"Critical": 2, "High": 8, "Medium": 24, "Low": 72}
-        sla_hours = hours_map.get(classification["priority"], 72)
-        sla_breach_dt = datetime.datetime.utcnow() + datetime.timedelta(hours=sla_hours)
-
-        ticket_response_dict = {
-            "ticket_id": str(uuid.uuid4()),
-            "summary": summary,
-            "category": classification["category"],
-            "subcategory": classification["subcategory"],
-            "priority": classification["priority"],
-            "auto_resolve": classification["auto_resolve"],
-            "assigned_team": classification["assigned_team"],
-            "entities": [e for e in entities],
-            "duplicate_ticket": dup_result,
-            "confidence": classification["confidence"],
-            "needs_review": classification["confidence"] < confidence_threshold,
-            "reasoning": reasoning,
-            "decision_factors": decision_factors,
-            "image_description": gemini_analysis["image_description"],
-            "ocr_text": gemini_analysis["ocr_text"],
-            "image_url": request_body.image_url,
-            "highlights": entities,
-            "timeline": timeline,
-            "env_metadata": env_metadata,
-            "sla_breach_at": sla_breach_dt.isoformat() + "Z"
-        }
-
-        # 6. Final Result
-        yield f"data: {json.dumps({'step': 'done', 'result': jsonable_encoder(ticket_response_dict)})}\n\n"
-
-    return StreamingResponse(event_generator(), media_type="text/event-stream")
-
-@app.post("/ai/analyze_ticket/legacy")
-async def legacy_analyze_and_save(request_body: TicketRequest):
-    """
-    BACKWARD COMPATIBILITY: Strictly performs analysis only. 
-    Does NOT persist to DB to avoid foreign key violations.
-    """
-    return await analyze_only(request_body)
-
-@app.post("/ai/analyze-v2")
-async def analyze_ticket_v2(request: TicketRequest):
-    text = request.text
-    try:
-        prediction = classifier_v2.predict(text)
-        return {
-            "status": "success",
-            "category": prediction["category"]["prediction"],
-            "subcategory": prediction["sub_category"]["prediction"],
-            "priority": prediction["priority"]["prediction"],
-            "auto_resolve": prediction["auto_resolve"]["prediction"].lower() == "true",
-            "assigned_team": prediction["assigned_team"]["prediction"],
-            "confidence": prediction["category"]["confidence"]
-        }
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-# ---------------------------------------------------------------------------
-# Clean cookie-based Supabase Auth endpoints for /auth/me backward-compatibility
-# ---------------------------------------------------------------------------
-ACCESS_COOKIE = "access_token"
-REFRESH_COOKIE = "refresh_token"
-ACCESS_MAX_AGE = 60 * 60
-REFRESH_MAX_AGE = 60 * 60 * 24 * 7
-
-def _cookie_kwargs() -> dict:
-    secure = os.getenv("ENV", "production").lower() != "development"
-    return {
-        "httponly": True,
-        "secure": secure,
-        "samesite": "strict",
-        "path": "/",
-    }
-
-def extract_token(request: Request) -> str | None:
-    cookie_token = request.cookies.get(ACCESS_COOKIE)
-    if cookie_token:
-        return cookie_token
-    auth = request.headers.get("authorization") or request.headers.get("Authorization")
-    if auth and auth.lower().startswith("bearer "):
-        return auth.split(" ", 1)[1].strip() or None
-    return None
-
-def _set_session_cookies(response: Response, session) -> None:
-    if not session or not getattr(session, "access_token", None):
-        return
-    response.set_cookie(
-        ACCESS_COOKIE,
-        session.access_token,
-        max_age=ACCESS_MAX_AGE,
-        **_cookie_kwargs(),
-    )
-    refresh = getattr(session, "refresh_token", None)
-    if refresh:
-        response.set_cookie(
-            REFRESH_COOKIE,
-            refresh,
-            max_age=REFRESH_MAX_AGE,
-            **_cookie_kwargs(),
-        )
-
-def _clear_session_cookies(response: Response) -> None:
-    kwargs = _cookie_kwargs()
-    response.delete_cookie(ACCESS_COOKIE, path=kwargs["path"])
-    response.delete_cookie(REFRESH_COOKIE, path=kwargs["path"])
-
-async def get_current_user(request: Request) -> dict:
-    token = extract_token(request)
-    if not token:
-        raise HTTPException(status_code=401, detail="Not authenticated")
-    if not supabase:
-        raise HTTPException(status_code=503, detail="Database connection offline")
-    try:
-        result = supabase.auth.get_user(token)
-    except Exception as exc:
-        raise HTTPException(
-            status_code=401,
-            detail=f"Invalid session: {exc}",
-        ) from exc
-    user = getattr(result, "user", None) or (result.get("user") if isinstance(result, dict) else None)
-    if not user:
-        raise HTTPException(status_code=401, detail="Invalid session")
-    if hasattr(user, "model_dump"):
-        return user.model_dump()
-    if hasattr(user, "dict"):
-        return user.dict()
-    return dict(user)
-
-class LoginBody(BaseModel):
-    email: str
-    password: str
-
-class SignupBody(BaseModel):
-    email: str
-    password: str
-    full_name: str | None = None
-    role: str | None = "user"
-    company: str | None = None
-
-@app.post("/auth/login")
-async def auth_login(body: LoginBody, response: Response):
-    if not supabase:
-        raise HTTPException(status_code=503, detail="Database connection offline")
-    try:
-        result = supabase.auth.sign_in_with_password(
-            {"email": body.email, "password": body.password}
-        )
-    except Exception as exc:
-        raise HTTPException(status_code=401, detail=str(exc)) from exc
-
-    session = getattr(result, "session", None)
-    user = getattr(result, "user", None)
-    if not session or not user:
-        raise HTTPException(status_code=401, detail="Invalid credentials")
-
-    _set_session_cookies(response, session)
-    user_payload = user.model_dump() if hasattr(user, "model_dump") else dict(user)
-    return {"user": user_payload, "message": "Session cookies set"}
-
-@app.post("/auth/signup")
-async def auth_signup(body: SignupBody, response: Response):
-    if not supabase:
-        raise HTTPException(status_code=503, detail="Database connection offline")
-    metadata = {}
-    if body.full_name:
-        metadata["full_name"] = body.full_name
-    if body.role:
-        metadata["role"] = body.role
-    if body.company:
-        metadata["company"] = body.company
-
-    try:
-        result = supabase.auth.sign_up(
-            {
-                "email": body.email,
-                "password": body.password,
-                "options": {"data": metadata} if metadata else {},
-            }
-        )
-    except Exception as exc:
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    session = getattr(result, "session", None)
-    user = getattr(result, "user", None)
-    if session:
-        _set_session_cookies(response, session)
-    user_payload = user.model_dump() if user and hasattr(user, "model_dump") else None
-    return {"user": user_payload, "message": "Signup complete"}
-
-@app.post("/auth/logout")
-async def auth_logout(response: Response):
-    _clear_session_cookies(response)
-    return {"ok": True}
-
-@app.get("/auth/me")
-async def auth_me(user: dict = Depends(get_current_user)):
-    return {"user": user}
