+++ backend/routers/auth.py
@@ -1,0 +1,135 @@
+
+import os
+from fastapi import APIRouter, Depends, HTTPException, Request, Response
+from backend.dependencies import supabase
+from backend.models import LoginBody, SignupBody
+
+router = APIRouter(prefix="/auth", tags=["auth"])
+ACCESS_COOKIE = "access_token"
+REFRESH_COOKIE = "refresh_token"
+ACCESS_MAX_AGE = 60 * 60
+REFRESH_MAX_AGE = 60 * 60 * 24 * 7
+
+def _cookie_kwargs() -> dict:
+    secure = os.getenv("ENV", "production").lower() != "development"
+    return {
+        "httponly": True,
+        "secure": secure,
+        "samesite": "strict",
+        "path": "/",
+    }
+
+def extract_token(request: Request) -> str | None:
+    cookie_token = request.cookies.get(ACCESS_COOKIE)
+    if cookie_token:
+        return cookie_token
+    auth = request.headers.get("authorization") or request.headers.get("Authorization")
+    if auth and auth.lower().startswith("bearer "):
+        return auth.split(" ", 1)[1].strip() or None
+    return None
+
+def _set_session_cookies(response: Response, session) -> None:
+    if not session or not getattr(session, "access_token", None):
+        return
+    response.set_cookie(
+        ACCESS_COOKIE,
+        session.access_token,
+        max_age=ACCESS_MAX_AGE,
+        **_cookie_kwargs(),
+    )
+    refresh = getattr(session, "refresh_token", None)
+    if refresh:
+        response.set_cookie(
+            REFRESH_COOKIE,
+            refresh,
+            max_age=REFRESH_MAX_AGE,
+            **_cookie_kwargs(),
+        )
+
+def _clear_session_cookies(response: Response) -> None:
+    kwargs = _cookie_kwargs()
+    response.delete_cookie(ACCESS_COOKIE, path=kwargs["path"])
+    response.delete_cookie(REFRESH_COOKIE, path=kwargs["path"])
+
+async def get_current_user(request: Request) -> dict:
+    token = extract_token(request)
+    if not token:
+        raise HTTPException(status_code=401, detail="Not authenticated")
+    if not supabase:
+        raise HTTPException(status_code=503, detail="Database connection offline")
+    try:
+        result = supabase.auth.get_user(token)
+    except Exception as exc:
+        raise HTTPException(
+            status_code=401,
+            detail=f"Invalid session: {exc}",
+        ) from exc
+    user = getattr(result, "user", None) or (result.get("user") if isinstance(result, dict) else None)
+    if not user:
+        raise HTTPException(status_code=401, detail="Invalid session")
+    if hasattr(user, "model_dump"):
+        return user.model_dump()
+    if hasattr(user, "dict"):
+        return user.dict()
+    return dict(user)
+
+
+@router.post("/login")
+async def auth_login(body: LoginBody, response: Response):
+    if not supabase:
+        raise HTTPException(status_code=503, detail="Database connection offline")
+    try:
+        result = supabase.auth.sign_in_with_password(
+            {"email": body.email, "password": body.password}
+        )
+    except Exception as exc:
+        raise HTTPException(status_code=401, detail=str(exc)) from exc
+
+    session = getattr(result, "session", None)
+    user = getattr(result, "user", None)
+    if not session or not user:
+        raise HTTPException(status_code=401, detail="Invalid credentials")
+
+    _set_session_cookies(response, session)
+    user_payload = user.model_dump() if hasattr(user, "model_dump") else dict(user)
+    return {"user": user_payload, "message": "Session cookies set"}
+
+@router.post("/signup")
+async def auth_signup(body: SignupBody, response: Response):
+    if not supabase:
+        raise HTTPException(status_code=503, detail="Database connection offline")
+    metadata = {}
+    if body.full_name:
+        metadata["full_name"] = body.full_name
+    if body.role:
+        metadata["role"] = body.role
+    if body.company:
+        metadata["company"] = body.company
+
+    try:
+        result = supabase.auth.sign_up(
+            {
+                "email": body.email,
+                "password": body.password,
+                "options": {"data": metadata} if metadata else {},
+            }
+        )
+    except Exception as exc:
+        raise HTTPException(status_code=400, detail=str(exc)) from exc
+
+    session = getattr(result, "session", None)
+    user = getattr(result, "user", None)
+    if session:
+        _set_session_cookies(response, session)
+    user_payload = user.model_dump() if user and hasattr(user, "model_dump") else None
+    return {"user": user_payload, "message": "Signup complete"}
+
+@router.post("/logout")
+async def auth_logout(response: Response):
+    _clear_session_cookies(response)
+    return {"ok": True}
+
+@router.get("/me")
+async def auth_me(user: dict = Depends(get_current_user)):
+    return {"user": user}
+
